로그인, 회원가입 구현

+++ mysite/blog/views.py
@@ -1,4 +1,6 @@
 from django.shortcuts import render, get_object_or_404, redirect
+from django.contrib.auth.decorators import login_required
+from django.http import HttpResponseForbidden
 from .models import Post
 from .forms import PostForm
 
@@ -13,12 +15,13 @@ def detail(request, post_id):
     return render(request, 'blog/detail.html', {'post': post})
 
 
+@login_required
 def create(request):
     if request.method == 'POST':
         form = PostForm(request.POST)
         if form.is_valid():
-            post = form.save(commit=False)  # DB 저장 전 객체만 생성
-            post.author = request.user       # 현재 로그인 사용자를 작성자로
+            post = form.save(commit=False)
+            post.author = request.user
             post.save()
             return redirect('detail', post_id=post.id)
     else:
@@ -26,20 +29,26 @@ def create(request):
     return render(request, 'blog/form.html', {'form': form, 'title': '새 글 작성'})
 
 
+@login_required
 def update(request, post_id):
     post = get_object_or_404(Post, id=post_id)
+    if post.author != request.user:
+        return HttpResponseForbidden("본인의 글만 수정할 수 있습니다.")
     if request.method == 'POST':
-        form = PostForm(request.POST, instance=post)  # 기존 데이터에 덮어쓰기
+        form = PostForm(request.POST, instance=post)
         if form.is_valid():
             form.save()
             return redirect('detail', post_id=post.id)
     else:
-        form = PostForm(instance=post)  # 기존 데이터를 폼에 채워서 보여줌
+        form = PostForm(instance=post)
     return render(request, 'blog/form.html', {'form': form, 'title': '글 수정'})
 
 
+@login_required
 def delete(request, post_id):
     post = get_object_or_404(Post, id=post_id)
+    if post.author != request.user:
+        return HttpResponseForbidden("본인의 글만 삭제할 수 있습니다.")
     if request.method == 'POST':
         post.delete()
         return redirect('index')
